Did a lot of cleanup in the test classes
All the tests classes run with their own sql server database and the
databases are dostroyed when testing is compete. Not all the tests
reauired an IAsyncLifetime to handle databse cleanup.

Currently there are only two test classes that need to have the
IAsyncLifetime and need to cleanup after themselves.